extract reusable types for data grid components
Moves type definitions for data grid columns, sorting, and pagination
to a dedicated file to improve modularity and reusability.

Simplifies the use-data-grid logic by importing shared types.
